% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.r
\name{fibonacci}
\alias{fibonacci}
\title{fibonacci}
\usage{
fibonacci(n = 5, save = FALSE, print = FALSE)
}
\arguments{
\item{n}{The number of Fibonacci numbers to generate.}

\item{save}{Logical. If TRUE, the generated numbers are saved in a file.}

\item{print}{Logical. If TRUE, the generated numbers are printed to the console.}

\item{fibonacci}{Generates n Fibonacci numbers}
}
\value{
A sequence of fibonacci numbers or a csv file
}
\description{
Generates n Fibonacci numbers as a sequence
}
\examples{
fib1 <- fibonacci(n=50,save=FALSE,print=FALSE)
fibonacci(n=50,save=FALSE,print=TRUE)
}
