% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leibniz π.r
\name{pi}
\alias{pi}
\title{pi}
\arguments{
\item{pi}{A Liebniz approximation of pi. Generates Pi to varying precision. Default iterations is 10e6.}
}
\value{
value of Pi to varying precision
}
\description{
pi
}
\examples{
pi <- pi(digits=50);
}
