% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.R
\name{zero}
\alias{zero}
\title{zero
prints zero if you need a zero anytime}
\usage{
zero(printing = T)
}
\arguments{
\item{printing}{Option to return or print the value to console}

\item{zero}{Just zero}
}
\value{
zero
}
\description{
zero
prints zero if you need a zero anytime
}
\examples{
zero <- zero();
}
