% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fibonacci.r
\name{fibonacci}
\alias{fibonacci}
\title{fibonacci}
\arguments{
\item{fibonacci}{Generates fibonacci}
}
\value{
A sequence of fibonacci numbers or a csv file
}
\description{
fibonacci
}
\examples{
fib1 <- fibonacci(n=50);
fibonacci(n=50,save=T,print=T)
}
