\name{Pi}
\alias{Pi}

\title{
Pi
}
\description{
Prints the value of pi to any number of digits
}
\usage{
pi(x)
}

\description{

Leibniz π/4 formula

We use Leibniz (π/4) formula. It states that

The series is convergent and continuous (Ref. Leibniz's test for convergence). The series converges to π(after multiplying by 4) when we take a sum of infinite terms.
Since the convergence to π is very slow, the algorithm requires a summation of large terms to get an accurate value of π for small number of digits.
In other terms, the algorithm is super slow to calculate the value of pi.
}

\author{
Ohnoimded}
\note{
Performance depends on the hardware. Will update with better code}


\examples{print(pi(digits=13))

}
