% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leibniz π.r
\name{pi}
\alias{pi}
\title{pi}
\usage{
pi(precision = 1e+09, digits = 2000)
}
\arguments{
\item{pi}{A Liebniz approximation of pi}
}
\value{
value of Pi to varying precision
}
\description{
Generates Pi to varying precision. Default iterations is 10e6.
}
\examples{
fib1 <- pi(digits=50);
}
