% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.r
\name{pi}
\alias{pi}
\title{pi}
\usage{
pi(precision = 10000, digits = 2000)
}
\arguments{
\item{precision}{Number of iterations to get precise value}

\item{digits}{Number of digits after the decimal point}

\item{pi}{A Leibniz approximation of pi. Generates Pi to varying precision. Default iterations is 10e3.}
}
\value{
Value of pi at varying number of digits
}
\description{
Prints the value of pi
}
\examples{
pi <- pi(digits=50,precision = 10e7);
}
